json: Support timeout
It's a good idea to set a timeout for HTTP requests. This is particularly helpful when we use this package via analytics.js and localstorage-retry. Setting a timeout will help avoid resource contention when the network is flaky.

The timeout must be passed as an option. A timeout of zero indicates no timeout. This is a breaking API change, so I'll be releasing this as a new major version.

Since this is a breaking change, I removed support for optional arguments to make the code easier to read. At least in a.js, We use all the options anyway.